+++ src/utils/type.ts
@@ -1,0 +1,6 @@
+type MenuCategory = "espresso" | "frappuchino" | "blended" | "teavana" | "dessert";
+
+
+export {
+  MenuCategory
+}
