+++ tsconfig.json
@@ -3,6 +3,7 @@
     "allowJs": true,
     "target": "es5",
     "outDir": "./build",
+    "module": "ES2015",
     "moduleResolution": "Node",
     "lib": ["ES2015", "DOM", "DOM.Iterable"],
   },
